[ContainersPreview] Spin off removal/consumption operations from RangeReplaceableContainer into a new DrainableContainer protocol
We have types that can usefully conform to `DrainableContainer` but not the full `RangeReplaceableContainer`; `OrderedSet` is one obvious example.
