+++ Sources/ContainersPreview/CMakeLists.txt
@@ -38,6 +38,7 @@ target_sources(${module_name} PRIVATE
   "Protocols/Container/Container+Filter.swift"
   "Protocols/Container/Container+SpanwiseZip.swift"
   "Protocols/Container/ContainerIterator.swift"
+  "Protocols/Container/DrainableContainer.swift"
   "Protocols/Container/DynamicContainer.swift"
   "Protocols/Container/MutableContainer.swift"
   "Protocols/Container/PermutableContainer.swift"
